# Apps/swagger.json
{
  "swagger": "2.0",
  "basePath": "/",
  "paths": {
...
    "/Users/add_user": {
      "post": {
        "responses": {
          "400": {
            "description": "Missing required fields"
          },
          "201": {
            "description": "User created successfully"
          }
        },
        "summary": "Create a new user",
        "operationId": "add_user",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
...
    "/users/delete_user/{user_id}": {
      "parameters": [
        {
          "name": "user_id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ],
      "delete": {
        "responses": {
          "404": {
            "description": "User not found"
          },
          "200": {
            "description": "User deleted successfully"
          }
        },
        "summary": "Delete a user by ID",
        "operationId": "delete_user",
        "tags": [
          "users"
        ]
      }
    },
    "/users/search_users": {
      "get": {
        "responses": {
          "404": {
            "description": "No users found"
          },
          "400": {
            "description": "Bad request"
          },
          "200": {
            "description": "Users found"
          }
        },
        "summary": "Search users by name or email",
        "operationId": "search_users",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Search_User"
            }
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
    "/users/{user_id}": {
      "parameters": [
        {
          "in": "path",
          "description": "The user identifier",
          "name": "user_id",
          "required": true,
          "type": "string"
        }
      ],
      "get": {
        "responses": {
          "404": {
            "description": "User not found"
          },
          "200": {
            "description": "Success"
          }
        },
        "summary": "Fetch a user by ID",
        "operationId": "get_user",
        "tags": [
          "users"
        ]
      },
      "put": {
        "responses": {
          "400": {
            "description": "Bad request"
          },
          "200": {
            "description": "User updated successfully"
          }
        },
        "summary": "Update a user by ID",
        "operationId": "update_user",
        "parameters": [
          {
            "name": "payload",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UpdateUser"
            }
          }
        ],
        "tags": [
          "users"
        ]
      }
    }
  },
  "info": {
    "title": "User Management API",
    "version": "1.0",
    "description": "A simple User Management API"
  },
  "produces": [
    "application/json"
  ],
  "consumes": [
    "application/json"
  ],
...
  "definitions": {
    "UpdateUser": {
      "properties": {
        "UserName": {
          "type": "string",
          "description": "The user name"
        },
        "Email": {
          "type": "string",
          "description": "The user email"
        },
        "Phone": {
          "type": "string",
          "description": "The user phone"
        },
        "Address": {
          "type": "string",
          "description": "The user address"
        },
        "Status": {
          "type": "string",
          "description": "The user status"
        }
      },
      "type": "object"
    },
    "Search_User": {
      "properties": {
        "name": {
          "type": "string",
          "description": "The name to search for"
        },
        "email": {
          "type": "string",
          "description": "The email to search for"
        }
      },
      "type": "object"
    },
...
    "User": {
      "required": [
        "Name",
        "UserId"
      ],
      "properties": {
        "UserId": {
          "type": "string",
          "description": "The user ID"
        },
        "Name": {
          "type": "string",
          "description": "The user name"
        },
        "Email": {
          "type": "string",
          "description": "The user email",
          "default": "unknown@example.com"
        },
        "Status": {
          "type": "string",
          "description": "The user status",
          "default": "active"
        },
        "Preferences": {
          "type": "object",
          "description": "The user preferences",
          "default": {
            "theme": "light",
            "notifications": true
          }
        },
        "CreatedAt": {
          "type": "string",
          "description": "The user creation timestamp",
          "default": "2025-04-26T22:16:00.503739"
        }
      },
      "type": "object"
    },
...
  },
...
}